Fix Tablet and Mobile view for EE
